requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3